requirements.txt
greenlet==2.0.1
numpy==1.23.5
pandas==1.5.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
